moving around provider unit tests to proper files
This is a post provider refactor exercise and gets things ready
for predicate functions unit tests
